refactor: extract app shell sidebar
