+++ src/components/app-shell/app-shell-types.ts
@@ -1,0 +1,1 @@
+export type AppShellRoute = "archived" | "billing" | "chat" | "connections" | "organizations" | "skills" | "settings"
